refactor: share code
