+++ packages/article/src/article-lead-asset/article-lead-asset.web.js
@@ -9,20 +9,7 @@ import {
   LeadAsset,
   LeadAssetCaptionContainer
 } from "../styles/article-body/responsive";
-
-const AspectRatioContainer = ({ aspectRatio, children }) => {
-  const [ratioWidth, ratioHeight] = aspectRatio.split(":");
-  const aspectRatioPercent = ratioHeight / ratioWidth * 100;
-  return (
-    <View
-      style={{ position: "relative", paddingBottom: `${aspectRatioPercent}%` }}
-    >
-      <View style={{ position: "absolute", width: "100%", height: "100%" }}>
-        {children}
-      </View>
-    </View>
-  );
-};
+import AspectRatioContainer from "../media-aspect-ratio";
 
 AspectRatioContainer.propTypes = {
   aspectRatio: PropTypes.string.isRequired,
