+++ packages/article/src/media-aspect-ratio.js
